Set an undefined default_currency to "" instead of None as it's a CharField
We'll have to revisit this to decide if 'XXX' is a better default or not
but this fixes the errors in the test suite as it is.

If we were passed Money, use it's currency to set the default.

Here we also pass all other values through to the super class so we
don't have to keep adding kwargs as the Django API changes.